Update App.js and Context
Remove all calls to MLS and Licensing data; the data pools were restructured and not longer require this separation of data.

This data will be altered in the future on the backend to support real world data.

Updated Context to pass all necessary functions to child components in the tree.