Continued Dev
Making progress in development of generating the agent profile.

Not all data is being pushed from the server. For some reason columns are not coming through even  though the select option has been defined.

It is only pulling the brokerage name; not other details for the agent and/or for the brokerage.

Updated the Note component to make the request to the database. Pulling all notes; but need to pull the notes only assigned to the agent that is pulled.