Debugging Notes
Notes component does not rerender following submission of a new note.

The new note is created in the database but the page requires a refresh to redisplay the notes.

A way to solve this could be to return the agents along with the notes on the backend instead of making multiple requests.

Exploring this option.