Updated App.js
Refactored followToggle to the app head as state needs to be accessed by multiple components.

button is still not updating properly due to the order of functions. The check to see if an agent is followed after clicking follow happens before the POST request to the server and before the state is updated.